Yet more progress with dynamic Svg
